Extract md conversion code from browser code for easier cli usage
